requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-socks==0.11.0
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
mistralai==2.0.4
openai==2.28.0
opentelemetry-api==1.39.1
opentelemetry-semantic-conventions==0.60b1
phidata==2.7.10
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-socks==2.8.1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
tomli==2.4.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.0
